ignore even more Apple Clang shorten-64-to-32 warnings from ITK

+++ modules/Uncertainty/iAEntropy.cpp
@@ -36,7 +36,7 @@ void entropy(iAFilter* filter, QVariantMap const & parameters)
 	auto entropyFilter = EntropyFilter::New();
 	for (size_t i = 0; i < filter->inputCount(); ++i)
 	{
-		entropyFilter->SetInput(i, dynamic_cast<InputImageType*>(filter->imageInput(i)->itkImage()));
+		entropyFilter->SetInput(static_cast<int>(i), dynamic_cast<InputImageType*>(filter->imageInput(i)->itkImage()));
 	}
 	entropyFilter->SetNormalize(parameters["Normalize"].toBool());
 	entropyFilter->Update();
